weather-prediction-streamlit-frontend-main/requirements.txt
streamlit>=1.29.0
pandas>=2.1.3
matplotlib>=3.8.2
google-cloud-bigquery>=3.13.0
db_dtypes>=1.1.1